client/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
./def_proto-0.0.1-py3-none-any.whl
certifi==2021.5.30
charset-normalizer==2.0.4; python_version >= '3'
grpcio==1.39.0
idna==3.2; python_version >= '3'
requests==2.26.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
